requirements.txt
openai
pandas
requests
numpy
python-dotenv